==> 06_MixVsSingleCompounds.Rmd <==
---
title: "MixVsSingleCompunds"
output:
    html_document:
        toc: true
        toc_float: true
        collapsed: true
        theme: 'simplex'
        highlight: 'espresso'
        code_folding: hide
---

### Data loading

```{r, warning=FALSE, message=FALSE}
suppressPackageStartupMessages({
  library(SummarizedExperiment)  
  library(SEtools)
  library(edgeR)
  library(DT)
  library(pheatmap)
  library(plotly)
  library(dplyr)
  library(sva)
  library(ggrepel)
})
source("Functions/EDC_Functions.R")
source("Functions/CriFormatted.R")
load("Data/AllSEcorrected.RData",verbose=T)
load("Data/HormonalGenes/HormonalGenes.RData",verbose = T)
load("Data/DEAfetSingleCompunds.RData", verbose=TRUE)
load("Data/DEAorgSingleCompunds.RData", verbose=TRUE)
```

### MixN vs single compounds Fetal
#### Fetal MixN DEGs are specific for MixN treated samples
```{r, warning=FALSE, message=FALSE}
fetal <- SEs$chronic.fetal[,which(SEs$chronic.fetal$EXPO=="CNT" | SEs$chronic.fetal$EXPO=="1X" | SEs$chronic.fetal$EXPO=="1000X" | SEs$chronic.fetal$EXPO=="T3"| SEs$chronic.fetal$EXPO=="BPA0.04X")]

degs.fe <- row.names(DEAs$chronic.fetal)[which((abs(DEAs$chronic.fetal$logFC.EXPO1X)>0.5 | abs(DEAs$chronic.fetal$logFC.EXPO1000X)>0.5) & DEAs$chronic.fetal$FDR<=0.05 & DEAs$chronic.fetal$logCPM>0)]

sehm(fetal, assayName = "corrected", degs.fe,  do.scale = T, show_rownames = F, anno_columns = c("EXPO2","MixtureBatch"), main="MixNDEGs")
```

#### Fetal T3 treated samples respond to T3 known targets and the DEGs are specific for T3 samples


```{r, warning=FALSE, message=FALSE, fig.height=7.5, fig.width=7.5}
sehm(fetal[,which(SEs$chronic.fetal$EXPO=="CNT" | SEs$chronic.fetal$EXPO=="T3")], assayName = "corrected", Thyroid,  do.scale = T, show_rownames = F, anno_columns = c("EXPO2","MixtureBatch"),  main="T3 target genes")

degs.fetT3 <- row.names(res.fetT3)[which(abs(res.fetT3$logFC)> 0.5 & res.fetT3$FDR<=0.05 & res.fetT3$logCPM > 0)]
sehm(fetal, assayName = "corrected", degs.fetT3,  do.scale = T, show_rownames = F, anno_columns = c("EXPO2","MixtureBatch"), main="T3DEGs")


T3 <- res.fetT3
T3$genes <- row.names(T3)

MixN <- DEAs$chronic.fetal
MixN$genes <- row.names(MixN)
MixN$logFC <- (MixN$logFC.EXPO1X + MixN$logFC.EXPO1000X)/2

Comp <- compareResultsFCNew(T3, MixN, FDRth=0.05, FCth=2^0.5, FDRceil=1e-10, logCPMth=0, FCceil=5, 
                              title='Fetal: Comparing T3 to MixN (mean FC)', geneLabel=TRUE, topLab=-30)
Comp$Scatter

#ggsave(Comp$Scatter, filename='Data/SingleCompounds/ScatterFetal_MixN_T3.pdf', width=10, height=10)
#ggsave(Comp$Scatter, filename='Data/SingleCompounds/ScatterFetal_MixN_T3.png', width=10, height=10)
```

```{r}
# Check overlap between MixN and Thyroid
table(Comp$AllRes$Status)

## Fisher Test Significant
SigBoth <- dplyr::filter(Comp$Significant, Comp$Significant$Status=='SignificantA&B')

Contingency <- matrix(c(dim(dplyr::filter(SigBoth, log2FCA < 0 & log2FCB < 0 ))[1], 
                          dim(dplyr::filter(SigBoth, log2FCA >= 0 & log2FCB < 0 ))[1],
                           dim(dplyr::filter(SigBoth, log2FCA < 0 & log2FCB >= 0 ))[1], 
                            dim(dplyr::filter(SigBoth, log2FCA >= 0 & log2FCB >= 0 ))[1]),
                        nrow=2, dimnames=list(c('DownA', 'UpA'), c('DownB', 'UpB')))
Contingency
```


#### Fetal BPA treated samples show specific DEGs
```{r, warning=FALSE, message=FALSE, fig.height=7.5, fig.width=7.5}
degs.fetBPA <- row.names(res.fetBPA)[which(abs(res.fetBPA$logFC)> 0.5 & res.fetBPA$FDR<=0.05 & res.fetBPA$logCPM > 0)]

sehm(fetal, assayName = "corrected", degs.fetBPA,  do.scale = T, show_rownames = F, anno_columns = c("EXPO2","MixtureBatch"), main="BPADEGs")

BPA <- res.fetBPA
BPA$genes <- row.names(BPA)

MixN <- DEAs$chronic.fetal
MixN$genes <- row.names(MixN)
MixN$logFC <- MixN$logFC.EXPO1X

Comp <- compareResultsFCNew(BPA, MixN, FDRth=0.05, FCth=2^0.5, FDRceil=1e-10, logCPMth=0, FCceil=4, 
                              title='Fetal: Comparing BPA to MixN (1X FC)', geneLabel=TRUE, topLab=-30)
Comp$Scatter

#ggsave(Comp$Scatter, filename='Data/SingleCompounds/ScatterFetal_MixN_BPA.pdf', width=10, height=10)
#ggsave(Comp$Scatter, filename='Data/SingleCompounds/ScatterFetal_MixN_BPA.png', width=10, height=10)
```

### MixN vs single compounds Organoids

#### Organoids MixN DEGs are specific for MixN treated samples
```{r, warning=FALSE, message=FALSE}
Organoids <- SEs$chronic.org[,which(SEs$chronic.org$EXPO=="CNT"|SEs$chronic.org$EXPO=="1X" | SEs$chronic.org$EXPO=="1000X" | SEs$chronic.org$EXPO=="T3"| SEs$chronic.org$EXPO=="BPA0.04X")]
degs.org <- row.names(DEAs$chronic.org)[which((abs(DEAs$chronic.org$logFC.EXPO1X)>0.5 | abs(DEAs$chronic.org$logFC.EXPO1000X)>0.5) & DEAs$chronic.org$FDR<=0.05 & DEAs$chronic.org$logCPM>0)]

sehm(Organoids, assayName = "corrected", degs.org,  do.scale = T, show_rownames = F, anno_columns = c("EXPO2","MixtureBatch"), main="MixNDEGs")
```

#### Organoids T3 treated samples respond to T3 known targets and the DEGs are specific for T3 samples


```{r, warning=FALSE, message=FALSE, fig.height=7.5, fig.width=7.5}
sehm(Organoids[,which(SEs$chronic.org$EXPO=="CNT" | SEs$chronic.org$EXPO=="T3")], assayName = "corrected", Thyroid,  do.scale = T, show_rownames = F, anno_columns = c("EXPO2","MixtureBatch"), main="T3 target genes")


degs.orgT3 <- row.names(res.orgT3)[which(abs(res.orgT3$logFC)> 0.5 & res.orgT3$FDR<=0.05 & res.orgT3$logCPM > 0)]

sehm(Organoids, assayName = "corrected", degs.orgT3,  do.scale = T, show_rownames = F, anno_columns = c("EXPO2","MixtureBatch"), main="T3DEGs")

T3 <- res.orgT3
T3$genes <- row.names(T3)

MixN <- DEAs$chronic.org
MixN$genes <- row.names(MixN)
MixN$logFC <- (MixN$logFC.EXPO1X + MixN$logFC.EXPO1000X)/2

Comp <- compareResultsFCNew(T3, MixN, FDRth=0.05, FCth=2^0.5, FDRceil=1e-10, logCPMth=0, FCceil=6, 
                              title='Comparing T3 to MixN (mean FC)', geneLabel=TRUE, topLab=-40)
Comp$Scatter

#ggsave(Comp$Scatter, filename='Data/SingleCompounds/ScatterOrg_MixN_T3.pdf', width=10, height=10)
#ggsave(Comp$Scatter, filename='Data/SingleCompounds/ScatterOrg_MixN_T3.png', width=10, height=10)
```

```{r}
# Check overlap between MixN and Thyroid
table(Comp$AllRes$Status)

## Contingency table on shared genes
SigBoth <- dplyr::filter(Comp$Significant, Comp$Significant$Status=='SignificantA&B')
ContingencyII <- matrix(c(dim(dplyr::filter(SigBoth, log2FCA < 0 & log2FCB < 0 ))[1], 
                          dim(dplyr::filter(SigBoth, log2FCA >= 0 & log2FCB < 0 ))[1],
                           dim(dplyr::filter(SigBoth, log2FCA < 0 & log2FCB >= 0 ))[1], 
                            dim(dplyr::filter(SigBoth, log2FCA >= 0 & log2FCB >= 0 ))[1]),
                        nrow=2, dimnames=list(c('DownA', 'UpA'), c('DownB', 'UpB')))
ContingencyII
```

```{r, collapse=TRUE}
prop.table(ContingencyII)

dplyr::filter(SigBoth, log2FCA < 0 & log2FCB > 0 ) %>% dplyr::pull(genes)
```


#### Organoids BPA treated samples show specific DEGs
```{r, warning=FALSE, message=FALSE, fig.height=7.5, fig.width=7.5}
degs.orgBPA <- row.names(res.orgBPA)[which(abs(res.orgBPA$logFC)> 0.5 & res.orgBPA$FDR<=0.05 & res.orgBPA$logCPM > 0)]

sehm(Organoids, assayName = "corrected", degs.orgBPA,  do.scale = T, show_rownames = F, anno_columns = c("EXPO2","MixtureBatch"),  main="BPADEGs")

BPA <- res.orgBPA
BPA$genes <- row.names(BPA)

MixN <- DEAs$chronic.org
MixN$genes <- row.names(MixN)
MixN$logFC <- MixN$logFC.EXPO1X

#Comp <- compareResultsFC(BPA, MixN, FDRth=0.05, FCth=2^0.5, FDRceil=1e-10, FCceil=8,logCPMth = 0, title='Comparing BPA to MixN 1X', geneLabel=TRUE)
#Comp$Scatter

Comp <- compareResultsFCNew(BPA, MixN, FDRth=0.05, FCth=2^0.5, FDRceil=1e-10, logCPMth=0, FCceil=4, 
                              title='Organoids: Comparing BPA to MixN (1X FC)', geneLabel=TRUE, topLab=-30)
Comp$Scatter

#ggsave(Comp$Scatter, filename='Data/SingleCompounds/ScatterOrg_MixN_BPA.pdf', width=10, height=10)
#ggsave(Comp$Scatter, filename='Data/SingleCompounds/ScatterOrg_MixN_BPA.png', width=10, height=10)
```


```{r}
# Check overlap between MixN and BPA
table(Comp$AllRes$Status)

## Contingency table on shared genes
SigBoth <- dplyr::filter(Comp$Significant, Comp$Significant$Status=='SignificantA&B')
ContingencyII <- matrix(c(dim(dplyr::filter(SigBoth, log2FCA < 0 & log2FCB < 0 ))[1], 
                          dim(dplyr::filter(SigBoth, log2FCA >= 0 & log2FCB < 0 ))[1],
                           dim(dplyr::filter(SigBoth, log2FCA < 0 & log2FCB >= 0 ))[1], 
                            dim(dplyr::filter(SigBoth, log2FCA >= 0 & log2FCB >= 0 ))[1]),
                        nrow=2, dimnames=list(c('DownA', 'UpA'), c('DownB', 'UpB')))
ContingencyII
```


### Data praparation

For details on data filtering, normalization, batch correction and differential expression analysis, see [here](./DataPreparation/SVAcorrectedObjects.html) and [here](./DataPreparation/DEAsingleCompunds.html) 

***

```{r child='footer.Rmd'}
```


```{r SaveSession}
sessionInfo()
```
