==> 08_HormonalPathways.Rmd <==
---
title: "HormonalPathways"
output:
    html_document:
        toc: true
        toc_float: true
        collapsed: true
        theme: 'simplex'
        highlight: 'espresso'
        code_folding: hide
---

### Data loading

```{r, warning=FALSE, message=FALSE}
suppressPackageStartupMessages({
  library(SummarizedExperiment)
  library(SEtools)
  library(edgeR)
  library(DT)
  library(pheatmap)
  library(plotly)
  library(dplyr)
  library(sva)
  library(msigdbr)
  library(reshape2)
  library(overlapper)
})
source("Functions/EDC_Functions.R")
source("Functions/CriFormatted.R")
load("Data/geneLengths.RData")
load("Data/AllSEcorrected.RData",verbose=T)
load("Data/TotalSumExp.RData", verbose = T)
```



## Overlap DEGs long list and hormonal genes
```{r, message=FALSE,collapse=TRUE, fig.height=6}
load("Data/HormonalGenes/HormonalGenes.RData",verbose = T)
load("Data/HormonalGenes/HormonalTFs/TF.RData", verbose = T)

uni <-rownames(filterGenes(Total))

degs.fe <- row.names(DEAs$chronic.fetal)[which(DEAs$chronic.fetal$FDR<=0.1)]
degs.org <- row.names(DEAs$chronic.org)[which(DEAs$chronic.org$FDR<=0.1 )]
degs <- list(Fetal=degs.fe, Organoids= degs.org, Both= union(degs.org,degs.fe))
TF <- list(Fetal=FetTF, Organoids= OrgTF, Both= union(OrgTF,FetTF))
hormonal <- list(Thyroid=Thyroid, Androgen=Androgen, Estrogen=Estrogen, Corticoid=Corticoid,Progesterone=Progesterone,PPAR=PPAR, Retinoic=Retinoic, All= union(Thyroid,union(Androgen, union(Estrogen,union(Corticoid, union(Progesterone,union(PPAR,Retinoic)))))))

m <- overlapper::multintersect(ll = degs, ll2 = hormonal, universe = uni,two.tailed = F)
dotplot.multintersect(m, sizeRange = c(0,15), th=0.05)
```


## Overlap TFs and hormonal genes

```{r, message=FALSE,collapse=TRUE,fig.height=6}
#load("Data/regulon_weighted.RData", verbose = T)
ag <- readRDS("Data/psychencode_reduced.rds")

col <- c("low"="grey", "high"="yellow")
#uni <- as.character(names(regulon))
uni <- as.character(unique(ag$Transcription_Factor))
m <- overlapper::multintersect(TF, hormonal, universe = uni,two.tailed = F)
dotplot.multintersect(m, sizeRange = c(0,15), th=0.05,colors = col)
```



## Heatmaps of hormonal DEGs
```{r, message=FALSE,collapse=TRUE}
SEs$chronic.org <- SEs$chronic.org[,which(SEs$chronic.org$EXPO=="CNT"|SEs$chronic.org$EXPO=="1X" | SEs$chronic.org$EXPO=="1000X")]
SEs$chronic.fetal <- SEs$chronic.fetal[,which(SEs$chronic.fetal$EXPO=="CNT"|SEs$chronic.fetal$EXPO=="1X" | SEs$chronic.fetal$EXPO=="1000X")]

SEs <- list(chronic.org=SEs$chronic.org, chronic.fetal=SEs$chronic.fetal)

merged <- mergeSEs(SEs, use.assays = "corrected", commonOnly=T)

HormonalAll <- union(Thyroid,union(Androgen, union(Estrogen,union(Corticoid, union(Progesterone,union(PPAR,Retinoic))))))
degs<- union(degs.fe,degs.org)
TF<- union(FetTF,OrgTF)

crossHm(SEs, genes = intersect(degs,HormonalAll), do.scale = T,assayName = "corrected",  show_rownames = F, anno_columns = c("EXPO2"), name = "Endocrine Relevant DEGs")

load("Data/HormonalGenes/HormonalDEGs/HormonalDEGs.RData", verbose = T)
```

#### Heatmaps of Thyroid DEGs
```{r, message=FALSE,collapse=TRUE}
crossHm(SEs, genes = ThyroidDEGs, do.scale = T, assayName = "corrected",show_rownames = F, anno_columns = c("EXPO2"), name = "Thyroid DEGs")
```

#### Heatmaps of Androgen DEGs
```{r, message=FALSE,collapse=TRUE}

crossHm(SEs, genes = AndrogenDEGs, do.scale = T, assayName = "corrected",show_rownames = F, anno_columns = c("EXPO2"), name = "Androgen DEGs")
```

#### Heatmaps of Estrogen DEGs
```{r, message=FALSE,collapse=TRUE}

crossHm(SEs, genes = EstrogenDEGs, do.scale = T, assayName = "corrected",show_rownames = F, anno_columns = c("EXPO2"), name = "Estrogen DEGs")
```

#### Heatmaps of Corticoid DEGs
```{r, message=FALSE,collapse=TRUE}

crossHm(SEs, genes = CorticoidDEGs, do.scale = T, assayName = "corrected",show_rownames = F, anno_columns = c("EXPO2"), name = "Corticoid DEGs")
```

#### Heatmaps of PPAR DEGs
```{r, message=FALSE,collapse=TRUE}

crossHm(SEs, genes = PPARDEGs, do.scale = T, assayName = "corrected",show_rownames = F, anno_columns = c("EXPO2"), name = "PPAR DEGs")
```

#### Heatmaps of Progesterone DEGs
```{r, message=FALSE,collapse=TRUE}

crossHm(SEs, genes = ProgesteroneDEGs, do.scale = T, assayName = "corrected",show_rownames = F, anno_columns = c("EXPO2"), name = "Progesterone DEGs")
```

#### Heatmaps of Retinoic DEGs
```{r, message=FALSE,collapse=TRUE}

crossHm(SEs, genes = RetinoicDEGs, do.scale = T, assayName = "corrected",show_rownames = F, anno_columns = c("EXPO2"), name = "Retinoic DEGs")

```

## Heatmaps of hormonal Psychencode Transciprion Factors
```{r, message=FALSE,collapse=TRUE}
load("Data/HormonalGenes/HormonalTFs/Psychencode/HormonalPsychencode.RData", verbose = T)


crossHm(SEs, genes = intersect(TF,HormonalAll), do.scale = T,assayName = "corrected",  show_rownames = F, anno_columns = c("EXPO2"), name = "Endocrine Relevant Transcription Factors")
```

#### Heatmaps of Thyroid TFs
```{r, message=FALSE,collapse=TRUE}
crossHm(SEs, genes = ThyroidTF, do.scale = T, assayName = "corrected",show_rownames = F, anno_columns = c("EXPO2"), name = "Thyroid Transcription Factors")
```

#### Heatmaps of Androgen TFs
```{r, message=FALSE,collapse=TRUE}

crossHm(SEs, genes = AndrogenTF, do.scale = T, assayName = "corrected",show_rownames = F, anno_columns = c("EXPO2"), name = "Androgen Transcription Factors")
```

#### Heatmaps of Estrogen TFs
```{r, message=FALSE,collapse=TRUE}

crossHm(SEs, genes = EstrogenTF, do.scale = T, assayName = "corrected",show_rownames = F, anno_columns = c("EXPO2"), name = "Estrogen Transcription Factors")
```

#### Heatmaps of Corticoid TFs
```{r, message=FALSE,collapse=TRUE}

crossHm(SEs, genes = CorticoidTF, do.scale = T, assayName = "corrected",show_rownames = F, anno_columns = c("EXPO2"), name = "Corticoid Transcription Factors")
```

#### Heatmaps of PPAR TFs
```{r, message=FALSE,collapse=TRUE}

crossHm(SEs, genes = PPARTF, do.scale = T, assayName = "corrected",show_rownames = F, anno_columns = c("EXPO2"), name = "PPAR Transcription Factors")
```

#### Heatmaps of Progesterone TFs
```{r, message=FALSE,collapse=TRUE}

crossHm(SEs, genes = ProgesteroneTF, do.scale = T, assayName = "corrected",show_rownames = F, anno_columns = c("EXPO2"), name = "Progesterone Transcription Factors")
```

#### Heatmaps of Retinoic TFs
```{r, message=FALSE,collapse=TRUE}

crossHm(SEs, genes = RetinoicTF, do.scale = T, assayName = "corrected",show_rownames = F, anno_columns = c("EXPO2"), name = "Retinoic Transcription Factors")

```


### Data praparation

For details on data filtering, normalization, batch correction and differential expression analysis, see [here](./DataPreparation/SVAcorrectedObjects.html) 

For details on the preparation of Hormonal pathways genes, see [here](./DataPreparation/HormonalPathway/DataPreparationHormonalGenes.html) 

***

```{r child='footer.Rmd'}
```



```{r SaveSession}
sessionInfo()
```
