==> 10_MixN_clustering.Rmd <==
---
title: "mixN clustering"
output:
    html_document:
        toc: true
        toc_float: true
        collapsed: true
        theme: 'simplex'
        highlight: 'espresso'
        code_folding: hide
---

### Data loading

```{r, warning=FALSE, message=FALSE}
suppressPackageStartupMessages({
  library(ComplexHeatmap)
  library(DT)
  library(ggplot2)
  library(cowplot)
  library(ggrepel)
  library(viper)
  library(limma)
  library(SummarizedExperiment)
  library(SEtools)
  library(igraph)
  library(overlapper)
})
theme_set(theme_cowplot())
source("Functions/EDC_Functions.R")
source("Functions/CriFormatted.R")
source("Functions/clustering.R")
source("Functions/clusterEnrichment.R")
```

```{r}
load("Data/AllSEcorrected.RData")
SEs <- lapply(SEs[1:3], FUN=function(x){
  x <- x[,x$EXPO!="VITC"]
  assays(x)$logFC <- assays(x)$corrected-rowMeans(assays(x)$corrected[,x$EXPO=="CNT"])
  x
})
DEAs <- DEAs[1:3]
# we get DEGs significant in at least one system
degs <- getDEGsFromRes(DEAs)
# we get 'cross-system degs'
tmp <- sapply(DEAs, FUN=function(x) x[unique(unlist(degs)),"FDR"])
q <- apply(tmp,1,weights=c(1,1,1),FUN=aggregation::lancaster)
names(q) <- unique(unlist(degs))
degs2 <- names(q)[q<0.05]
```

## Clusters of gene expression dyregulation between fetal and organoids
```{r, fig.height=5}
se <- mergeSEs(SEs[1:2], use.assays="logFC", do.scale = FALSE)
degs2 <- intersect(degs2,row.names(se))
se <- se[,se$EXPO %in% c("CNT","1X","1000X")]

# binarize
b <- sapply(split(seq_len(ncol(se)), paste(se$Dataset, se$EXPO)), FUN=function(i, lfct=0.1){
  x <- assays(se)$logFC[degs2,i,drop=FALSE]
  y <- sign(rowMedians(x))*(abs(rowMedians(x))>lfct)
  w <- apply(sign(x),1,FUN=function(x){
     max(as.numeric(table(x)))<(0.75*length(x))
  })
  y[w] <- y[w]/2
  y
})
row.names(b) <- degs2
set.seed(1234)
cl <- clusterWrapper(b[,grep("CNT",colnames(b),invert=TRUE)], 4:10, nstart=3)
rowData(se)$cluster <- factor(cl[["8"]]$cluster[row.names(se)])
cg <- split(names(cl[["8"]]$cluster), cl[["8"]]$cluster)

#cg2<-list()
#cg2$concordant <- cg[c(2,3,7,8)]
#cg2$discordant <- cg[c(1,4,6,5)]
#rowData(se)$cluster2 <-rowData(se)$cluster  
#rowData(se)$cluster2[which((rowData(se)$cluster==2)|(rowData(se)$cluster==3)|(rowData(se)$cluster==7)|(rowData(se)$cluster==8))] <- 1
#rowData(se)$cluster2[which((rowData(se)$cluster==1)|(rowData(se)$cluster==4)|(rowData(se)$cluster==6)|(rowData(se)$cluster==5))] <- 2


sehm(se, degs2, do.scale = FALSE, anno_columns = c("Dataset","EXPO2"), toporder = "cluster", anno_rows = "cluster", breaks=0.99,show_rownames = F,sortRowsOn = F,gaps_at = "Dataset")

#sehm(se, degs2, do.scale = FALSE, anno_columns = c("Dataset","EXPO2"), toporder = "cluster2", anno_rows = "cluster2", breaks=0.99,show_rownames = F,sortRowsOn = F)

# we isolate clusters that are coherent between the two systems
dc <- list(concordant=unlist(cg[c(2,3,7,8)]),
      up.both=unlist(cg[c(3,8)]),
      down.both=unlist(cg[c(2,7)]),
			discordant=unlist(cg[c(1)]),
			only_org=unlist(cg[c(4,6)]),
			only_fetal=unlist(cg[c(5)]))

save(dc, file = "Data/DEGsCluster.RData")
```

### Characterize the clusters

```{r, fig.height=7, fig.width=10}
sets <- getMsigSets("C5")
sets <- sets[grep("GO:BP",names(sets),fixed=TRUE)]
names(sets) <- gsub("C5:GO:BP:GO_","",names(sets),fixed=TRUE)
go <- clusterEnrichment(cg, sets=sets, universe=row.names(se))
#plotClusterEnrichment(go)

ag <- aggregate(assays(se)$logFC, by=list(cluster=rowData(se)$cluster), FUN=median)
row.names(ag) <- ag[,1]
se.ag <- SummarizedExperiment(list(logFC=ag[,-1]), colData=colData(se))
h1 <- sechm(se.ag, row.names(se.ag), do.scale = FALSE, anno_columns = c("Dataset","EXPO2"), breaks=1, show_rownames = F, width=unit(5,"cm"), sortRowsOn=NULL, cluster_rows=FALSE, row_title="Clusters", gaps_at="Dataset")
#h1
#m <- t(plotClusterEnrichment(go, returnMatrix=TRUE))
#colnames(m) <- breakStrings(tolower(gsub("_"," ",colnames(m))))
#h2 <- Heatmap(m, cluster_rows=FALSE, show_column_dend=FALSE, name="Enrichment")
#h1 + h2
```

```{r, fig.height=7, fig.width=10}
go2 <- goclassify(cg, sets)
#plot.goclassify(go2, n=16)
h1 + plot.goclassify(go2, n=16, transpose=TRUE, cluster_rows=FALSE, show_column_dend=FALSE)
```



## ASD genes

```{r, fig.width=8, fig.height=14}
load("Data/ASD.RData")
psychencode <- readRDS("Data/psychencode.DEGs.rds")

ASD <- lapply(ASD,FUN=function(x) gsub(" ","",x))
asd.degs <- intersect(unlist(dc[1]), unlist(ASD))
sehm(se, asd.degs, do.scale = FALSE, anno_columns = c("EXPO2","Dataset"), toporder = "cluster", show_rownames = TRUE, breaks=0.995, main="ASD-associated consistent DEGs")
```

### GO & ASD enrichments of consistent DEGs

```{r}
consistentDEGs <-unlist(dc[1])
oppositeDEGs <- unlist(dc[4])
m <- overlapper::multintersect(ll = list(consistentDEGs=consistentDEGs,oppositeDEGs=oppositeDEGs), ll2 = psychencode, universe = row.names(DEAs$chronic.fetal),two.tailed = F)

dotplot.multintersect(m, sizeRange = c(0,15), th=0.05)


go <- goseq.enrichment(row.names(se), unlist(dc[1]))
go[1:10,c(2,4,5,7)]

```


## Combining also acute exposure:

```{r, fig.width=12, fig.height=4}
se2 <- mergeSEs(SEs, use.assays="logFC", do.scale = FALSE)
sehm(se2[,grep("T3|BPA",colnames(se2),invert=TRUE)], intersect(degs$acute.fetal,degs2), do.scale = FALSE, anno_columns = c("Dataset","EXPO2"), breaks = 0.995, main="DEGs in acute exposure and at least one chronic system")
```

```{r, fig.width=12, fig.height=3}
sehm(se2[,grep("T3|BPA",colnames(se2),invert=TRUE)], intersect(degs$acute.fetal,unlist(dc[1])), do.scale = FALSE, anno_columns = c("Dataset","EXPO2"), breaks = 0.99, main=c("Consistent in chronic systems and\naltered upon acute exposure"))
```

### TF target enrichment among consistent DEGs

```{r}
load("Data/regulon_weighted.RData")
regulon <- lapply(regulon, FUN=function(x) intersect(row.names(se),names(x$tfmode)))
tf.en <- t(sapply(regulon, FUN=function(x){
  c( overlap.up=length(intersect(x, dc$up.both)),
     enrichment.up=log2(length(intersect(x, dc$up.both))/(length(dc$up.both)*length(x)/length(row.names(se)))),
     p.up=overlap.prob(x, dc$up.both, row.names(se)),
     overlap.down=length(intersect(x, dc$down.both)),
     enrichment.down=log2(length(intersect(x, dc$down.both))/(length(dc$down.both)*length(x)/length(row.names(se)))),
     p.down=overlap.prob(x, dc$down.both, row.names(se))
  )
}))
tf.en <- tf.en[which(rowMax(tf.en[,grep("overlap",colnames(tf.en))])>1),]
tf.en <- as.data.frame(tf.en[order(rowMin(tf.en[,grep("^p\\.",colnames(tf.en))])),])
tf.en$gene <- row.names(tf.en)
tf.en$q.up <- p.adjust(tf.en$p.up, "fdr")
tf.en$q.down <- p.adjust(tf.en$p.down, "fdr")
ggplot(tf.en, aes(enrichment.up, -log10(q.up), size=overlap.up)) + geom_point(alpha=0.3) + geom_text_repel(data=tf.en[which(tf.en$q.up<0.01),], mapping=aes(label=gene), show.legend = FALSE) + ggtitle("TF target enrichment among\nconsistently upregulated genes") + xlim(0,3)
```

Viper-based:

```{r}
se <- mergeSEs(SEs[1:2], use.assays="logFC", do.scale = FALSE)
se <- se[,se$EXPO %in% c("CNT","1X","1000X")]

load("Data/regulon_weighted.RData")
vi1 <- viper(assays(SEs$chronic.org[,SEs$chronic.org$EXPO %in% c("CNT","1X","1000X")])$corrected, regulon, verbose=FALSE)
vi2 <- viper(assays(SEs$chronic.fetal[,SEs$chronic.fetal$EXPO %in% c("CNT","1X","1000X")])$corrected, regulon, verbose=FALSE)
i <- intersect(row.names(vi1),row.names(vi2))
vi <- cbind(vi1[i,], vi2[i,])

se$EXPO <- droplevels(se$EXPO)
mm <- model.matrix(~se$Dataset+se$EXPO)
tf.en2 <- topTable(eBayes(lmFit(vi,mm)),coef = c("se$EXPO1X","se$EXPO1000X"), number=Inf)
head(tf.en2)
head(tf.en2[intersect(row.names(tf.en2),degs2),])
```

```{r child='footer.Rmd'}
```



```{r SaveSession}
sessionInfo()
```
